+++ backend/db/models/spot.js
@@ -96,7 +96,7 @@ module.exports = (sequelize, DataTypes) => {
       }
     },
     price: {
-      type:DataTypes.NUMERIC,
+      type:DataTypes.FLOAT,
       allowNull:false,
       validate: {
         min: 0
